Lean and mean global map of AudioContext ID -> worklet command channel
RwLock<Vec<_>> is a const fn so this saves a OnceLock. Since the path is
read-heavy anyway the RwLock is a better choice than a Mutex.

It does have a tiny memory leak (the Vec never cleans up) so probably
don't create a million of AudioContexts in one process please.